fix(spi,uart): Initialize eagerly
This commit changes the SPI and UART initialization so that all of the
possible SPIs and UARTs are enumerated and initialized together with the
AVR instead of on-demand.

Previously we were initializing those peripherals on the fist access -
when someone actually called e.g. `.uart0()` - and that's problematic
because it means that up until someone calls that function, we won't be
registered to that peripheral's `IRQ_OUTPUT` notification, meaning we
won't get its output sent _so far_.

This is best seen in the tests - before this commit the modified tests
would fail: we wouldn't have received `Ready!`, because the peripheral
wouldn't have been registered yet.

Closes https://github.com/Patryk27/avr-tester/issues/3